Separate the struct VendingMachineArchiver into a separate file
