+++ VendingMachineApp/VendingMachineApp/AppDelegate.swift
@@ -17,7 +17,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         do {
             try vendingMachine = VendingMachineArchiver.load()
         } catch {
-            vendingMachine = VendingMachineArchiver.setDefaultVendingMachine()
+            vendingMachine = VendingMachineArchiver.setDefault()
         }
         return true
     }
@@ -38,36 +38,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func applicationWillTerminate(_ application: UIApplication) {
         VendingMachineArchiver.archive(vendingMachine: vendingMachine)
-    }
-
-}
-
-struct VendingMachineArchiver {
-
-    enum ArchivingError: Error {
-        case noData
-        case notLoaded
-    }
-
-    static func archive(vendingMachine: VendingMachine?) {
-        guard let vendingMachine = vendingMachine else { return }
-        let vendingMachineEncoded = try? NSKeyedArchiver.archivedData(
-            withRootObject: vendingMachine,
-            requiringSecureCoding: false)
-        UserDefaults.standard.set(vendingMachineEncoded, forKey:"vendingMachine")
-    }
-
-    static func load() throws -> VendingMachine {
-        guard let data = UserDefaults.standard.data(forKey: "vendingMachine") else { throw ArchivingError.noData }
-        guard let vendingMachine = try NSKeyedUnarchiver
-            .unarchiveTopLevelObjectWithData(data) as? VendingMachine else { throw ArchivingError.notLoaded }
-        return vendingMachine
-    }
-
-    static func setDefaultVendingMachine() -> VendingMachine {
-        let money = Money()
-        let inventory = Inventory(list: [ObjectIdentifier: Pack]())
-        return VendingMachine.init(initialBalance: money, initialInventory: inventory)
     }
 
 }
